Structural Changes, Class Adding
I hate spaces in folder names, so I made a new sql folder, we should
discuss it.

Creating a class is now fully functional.

There were a few problems in the do.php file that I cleaned up. The
most up to date version is in the new sql folder.